+++ packages/components/src/index.ts
@@ -1,3 +1,4 @@
 export * from './components';
+export * from './molecules';
 export * from './hooks';
 export * from './test-utils';