+++ packages/utils/src/index.tsx
@@ -4,3 +4,4 @@ export * from './usePrevious';
 export * from './SearchSpecBuilder';
 export * from './sleep';
 export * from './generateShortId';
+export * from './getNormalisedError';