+++ packages/components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @batoanng/components
 
+## 2.0.4
+
+### Patch Changes
+
+- Introduce function NormalisedError
+- Updated dependencies
+  - @batoanng/utils@2.1.1
+
 ## 2.0.3
 
 ### Patch Changes