+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @batoanng/utils
 
+## 2.1.1
+
+### Patch Changes
+
+- Introduce function NormalisedError
+
 ## 2.1.0
 
 ### Minor Changes